extension: keep playback clock fast refreshable
